refactor : 로또와 당첨 domain 분리
