+++ docs/README.md
@@ -61,29 +61,31 @@
 
 # STEP3
 ## 로또(자동)
-### 로또 구입 금액
+### 로또
+#### 로또 구입 금액
 - [x] 로또 구매 금액을 생성한다.
   - [x] 숫자만 입력 가능하다.
   - [x] 1000의 배수이다.
   - [x] 1000 미만일 수 없다.
 - [x] 구매 가능 횟수를 계산한다.
-### 로또 번호 생성기
+#### 로또 번호 생성기
 - [ ] 구매 횟수를 속성으로 갖는다.
 - [x] 로또 번호들 리스트를 생성한다.
   - [x] 구매횟수만큼의 로또 번호 리스트를 생성한다.
   - [x] Collections.shuffle()을 활용한다.
-### 로또 번호
+#### 로또 번호
 - [x] 로또 번호를 생성한다.
   - [x] 45 초과의 수를 가질 수 없다.
   - [x] 1 미만의 수를 가질 수 없다.
-### 로또 번호들
+#### 로또 번호들
 - [ ] 로또 번호를 추가한다.  
   - [x] 6개 이상의 수를 추가 할 수 없다. 
   - [x] 중복된 수를 생성할 수 없다.
-### 로또 번호 출력기
+#### 로또 번호 출력기
 - [ ] 생성된 로또 번호를 출력한다.
   - Collections.sort() 메소드를 활용해 오름차순으로 출력한다.
-### 당첨 번호
+### 당첨
+#### 당첨 번호
 - [ ] 당첨 번호를 입력받는다.
   - [ ] 숫자만 입력 가능하다.
   - [ ] 1 미만의 수를 가질 수 없다.
@@ -91,20 +93,21 @@
   - [ ] 중복된 수를 입력할 수 없다.
 - [x] 일치 갯수를 판별한다.
   - [x] 로또 번호들을 받아 일치 갯수를 판별한다.
-### 당첨 금액 계산기
+#### 당첨 금액 계산기
 - [x] 일치 갯수 및 로또 갯수를 속성으로 갖는다.
 - [x] 일치 갯수별 담청 금액을 판별한다.
   - [x] 일치 갯수가 3개일 경우 5_000원
   - [x] 일치 갯수가 4개일 경우 50_000원
   - [x] 일치 갯수가 5개일 경우 1_500_000원
   - [x] 일치 갯수가 6개일 경우 2_000_000_000원
-### 총 당첨 금액 계산기
+#### 총 당첨 금액 계산기
 - [x] 당첨 금액 계산기 리스트를 속성으로 갖는다.
 - [x] 총 당첨 금액을 계산한다.
-### 수익률 계산기
+#### 수익률 계산기
 - [ ] 구매 구입 금액과 당첨 총 금액을 속성으로 갖는다.
 - [x] 수익률을 계산한다.
   - [x] 구매 구입 금액에서 당첨 금액을 차감한 후 구매 금액으로 나눈다.
-- [ ] 담청 통계를 출력한다.
-  - [ ] 일치 갯수별 담청 금액과 일치 로또 갯수를 출력한다.  
-  - [ ] 수익률을 출력한다.
+#### 당첨금액 출력기
+- [ ] 일치 갯수별 담청 금액과 일치 로또 갯수를 출력한다.
+#### 수익률 출력기
+- [ ] 수익률을 출력한다.
