+++ src/main/java/lotto/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.lotto.domain;
 
 import java.util.Objects;
