+++ src/main/java/lotto/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
