+++ src/main/java/lotto/winning/domain/TotalWinningMoneyCalculator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.winning.domain;
 
 import java.util.List;
