+++ src/main/java/lotto/winning/domain/WinningMoney.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.winning.domain;
 
 public enum WinningMoney {
