+++ src/test/java/lotto/lotto/domain/LottoPurchaseAmountFixture.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.lotto.domain;
+
+import lotto.lotto.domain.LottoPurchaseAmount;
 
 public class LottoPurchaseAmountFixture {
     public static LottoPurchaseAmount 로또구매금액_10_000() {
