+++ src/test/java/lotto/lotto/domain/LottoPurchaseAmountTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.lotto.domain;
 
+import lotto.lotto.domain.LottoPurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
