+++ src/test/java/lotto/lotto/domain/LottoTest.java
@@ -1,11 +1,12 @@
-package lotto;
+package lotto.lotto.domain;
 
+import lotto.lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static lotto.Lotto.MAXIMUM_NUMBER;
-import static lotto.Lotto.MINIMUM_NUMBER;
+import static lotto.lotto.domain.Lotto.MAXIMUM_NUMBER;
+import static lotto.lotto.domain.Lotto.MINIMUM_NUMBER;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
