+++ src/test/java/lotto/winning/domain/WinningMoneyCalculatorFixture.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.winning.domain;
+
+import lotto.winning.domain.WinningMoneyCalculator;
 
 import java.util.ArrayList;
 import java.util.List;
