+++ src/test/java/lotto/winning/domain/WinningNumberTest.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.winning.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.LottosFixture.로또번호123456;
-import static lotto.LottosFixture.로또번호123457;
-import static lotto.WinningNumberFixture.당첨번호123456;
+import static lotto.lotto.domain.LottosFixture.로또번호123456;
+import static lotto.lotto.domain.LottosFixture.로또번호123457;
+import static lotto.winning.domain.WinningNumberFixture.당첨번호123456;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("당첨번호")
